pyproject.toml
[project]
name = "shrutai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai-whisper>=20250625",
    "pyaudio>=0.2.14",
]
